requirements.txt
discord.py==1.7.3
pyfiglet==0.8.post1
Faker==12.0.0
python-decouple==3.5
requests
psutil
colorama
art
setuptools
PyYAML==6.0.1